Assistant checkpoint: Fix quantity handling for decimal values
Assistant generated file changes:
- app.py: Fix quantity conversion to handle decimal values
- models.py: Change quantity column to Float to support decimal quantities

---

User prompt:

im getting internal server error when i try to generate a pdf

+++ app.py
@@ -408,7 +408,7 @@ def create_invoice():
     for desc, price, qty in zip(descriptions, unit_prices, quantities):
         if desc and price and qty:
             unit_price = float(price)
-            quantity = int(qty)
+            quantity = float(qty)
             line_total = unit_price * quantity
 
             line = InvoiceLine(
